+++ Changes.md
@@ -1,3 +1,4 @@
+* #1270 - Add page displaying all known challenge descriptions
 * Allow the user to access database backups from the setup page
 * Correct links to download database backups
 * #1267 - On schedule upload filter columns that have data of the incorrect type